CSS counter-* should not eagerly evaluate calc() https://bugs.webkit.org/show_bug.cgi?id=279326
Reviewed by Darin Adler.

Replaces eager evaluation of calc() in counter-set, counter-increment, and
counter-reset, with storing the primitive value for resolution at style
building time.

Also took opportunity to move CSS List and CSS Counter Style consumers
to their own files.

* LayoutTests/imported/w3c/web-platform-tests/css/css-lists/parsing/counter-increment-valid-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-lists/parsing/counter-increment-valid.html:
* LayoutTests/imported/w3c/web-platform-tests/css/css-lists/parsing/counter-reset-valid-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-lists/parsing/counter-reset-valid.html:
* LayoutTests/imported/w3c/web-platform-tests/css/css-lists/parsing/counter-set-valid-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-lists/parsing/counter-set-valid.html:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/CSSCounterStyle.cpp:
* Source/WebCore/css/CSSCounterStyleRule.cpp:
* Source/WebCore/css/parser/CSSParserImpl.cpp:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+CounterStyles.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+CounterStyles.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Lists.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Lists.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp:
* Source/WebCore/css/parser/CSSPropertyParserHelpers.h:
* Source/WebCore/css/process-css-properties.py:

Canonical link: https://commits.webkit.org/283340@main
