+++ LayoutTests/imported/w3c/web-platform-tests/css/css-lists/parsing/counter-increment-valid-expected.txt
@@ -3,4 +3,7 @@ PASS e.style['counter-increment'] = "none" should set the property value
 PASS e.style['counter-increment'] = "chapter" should set the property value
 PASS e.style['counter-increment'] = "section -1" should set the property value
 PASS e.style['counter-increment'] = "first -1 second third 99" should set the property value
+PASS e.style['counter-increment'] = "section calc(1)" should set the property value
+PASS e.style['counter-increment'] = "section calc(-2.5)" should set the property value
+PASS e.style['counter-increment'] = "section calc(10 + (5 * sign(2cqw - 10px)))" should set the property value
