+++ LayoutTests/imported/w3c/web-platform-tests/css/css-lists/parsing/counter-increment-valid.html
@@ -15,6 +15,9 @@ test_valid_value('counter-increment', 'none');
 test_valid_value('counter-increment', 'chapter', 'chapter 1');
 test_valid_value('counter-increment', 'section -1');
 test_valid_value('counter-increment', 'first -1 second third 99', 'first -1 second 1 third 99');
+test_valid_value('counter-increment', 'section calc(1)');
+test_valid_value('counter-increment', 'section calc(-2.5)');
+test_valid_value('counter-increment', 'section calc(10 + (5 * sign(2cqw - 10px)))');
 </script>
 </body>
 </html>
