+++ LayoutTests/imported/w3c/web-platform-tests/css/css-lists/parsing/counter-reset-valid.html
@@ -21,6 +21,9 @@ test_valid_value('counter-reset', 'section -1');
 test_valid_value('counter-reset', 'reversed(section) -1');
 test_valid_value('counter-reset', 'first -1 second third 99', 'first -1 second 0 third 99');
 test_valid_value('counter-reset', 'reversed(first) -1 reversed(second) third 99', 'reversed(first) -1 reversed(second) third 99');
+test_valid_value('counter-reset', 'section calc(1)');
+test_valid_value('counter-reset', 'section calc(-2.5)');
+test_valid_value('counter-reset', 'section calc(10 + (5 * sign(2cqw - 10px)))');
 </script>
 </body>
 </html>
