+++ LayoutTests/imported/w3c/web-platform-tests/css/css-lists/parsing/counter-set-valid-expected.txt
@@ -3,4 +3,7 @@ PASS e.style['counter-set'] = "none" should set the property value
 PASS e.style['counter-set'] = "chapter" should set the property value
 PASS e.style['counter-set'] = "section -1" should set the property value
 PASS e.style['counter-set'] = "first -1 second third 99" should set the property value
+PASS e.style['counter-set'] = "section calc(1)" should set the property value
+PASS e.style['counter-set'] = "section calc(-2.5)" should set the property value
+PASS e.style['counter-set'] = "section calc(10 + (5 * sign(2cqw - 10px)))" should set the property value
