+++ Source/WebCore/css/parser/CSSPropertyParserConsumer+Filter.cpp
@@ -38,6 +38,7 @@
 #include "CSSPropertyParserConsumer+Percent.h"
 #include "CSSPropertyParserConsumer+URL.h"
 #include "CSSPropertyParserHelpers.h"
+#include "CSSToLengthConversionData.h"
 #include "CSSTokenizer.h"
 #include "CSSValueKeywords.h"
 #include "FilterOperations.h"
