+++ Source/WebCore/css/process-css-properties.py
@@ -3938,6 +3938,7 @@ class GenerateCSSPropertyParsing:
                     "CSSPropertyParserConsumer+Align.h",
                     "CSSPropertyParserConsumer+Angle.h",
                     "CSSPropertyParserConsumer+Color.h",
+                    "CSSPropertyParserConsumer+CounterStyles.h",
                     "CSSPropertyParserConsumer+Filter.h",
                     "CSSPropertyParserConsumer+Font.h",
                     "CSSPropertyParserConsumer+Grid.h",
@@ -3946,6 +3947,7 @@ class GenerateCSSPropertyParsing:
                     "CSSPropertyParserConsumer+Image.h",
                     "CSSPropertyParserConsumer+Length.h",
                     "CSSPropertyParserConsumer+List.h",
+                    "CSSPropertyParserConsumer+Lists.h",
                     "CSSPropertyParserConsumer+Number.h",
                     "CSSPropertyParserConsumer+Percent.h",
                     "CSSPropertyParserConsumer+Position.h",
